Benchmarking - Key-Value Store (#95)
This PR adds two different benchmarks. The first one makes requests with the same key to the servers, on the other hand the second one makes requests with different data to the servers.
Besides that, this PR also introduces this two different benchmarks in lua scripts and not only in locust scripts.